refactor(search): 💡 split functions into multiple files
